requirements.txt
absl-py==0.13.0
astor==0.8.1
atari-py==0.2.9
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
box2d-py==2.3.8
certifi==2021.5.30
charset-normalizer==2.0.12
cloudpickle==1.2.1
cycler==0.10.0
dataclasses==0.8
decorator==5.0.9
docopt==0.6.2
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.39.0
gym==0.15.7
h5py==2.10.0
idna==3.3
importlib-metadata==4.6.1
iniconfig==1.1.1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.1.1
mpi4py==3.0.3
numpy==1.18.5
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandas==1.1.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pipreqs==0.4.11
pluggy==0.13.1
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.8.0
pyglet==1.5.0
Pygments==2.9.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
pytz==2021.1
requests==2.27.1
scapy==2.4.5
scipy==1.5.4
seaborn==0.8.1
six==1.16.0
tensorboard==1.15.0
tensorflow==1.15.5
tensorflow-estimator==1.15.1
termcolor==1.1.0
timm==0.4.12
toml==0.10.2
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.2
traitlets==4.3.3
typing-extensions==3.10.0.0
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
yarg==0.1.9
zipp==3.5.0
